+++ Task.h
@@ -38,7 +38,7 @@ class Task {
 
   std::vector<RequestRow> _rows;
   std::vector<int> _indices;
-  std::atomic<size_t> _index = 0;
+  std::atomic<int> _index = 0;
   std::promise<void> _promise;
   const bool _diagnostics;
   Response _response;
